Completed preliminary trace routing
small tweeks made to schematic to reroute pot pins to other analog inputs.
still need to make ground and power planes with vias.
board is a little longer to correct for RJ45 placement.
switches were moved to make more room for the chips.
most or all caps and resistors were placed on the copper plane.